attendance_system-master/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
chardet==5.2.0
click==8.1.8
dotenv==0.9.9
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
qrcode==8.0
reportlab==4.3.1
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.24
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
qrcode~=8.0
reportlab~=4.3.1
pandas~=2.2.3
pillow~=11.1.0
openpyxl~=3.1.5


pdfplumber~=0.11.5